+++ build.sh
@@ -2,6 +2,8 @@
 
 # Run setup.
 . ./setup.sh
+. ./install-npm.sh
+. ./install-composer.sh
 
 # Loop through each wp-module in the plugin.
 for DIR in "$plugindir"/wp-modules/*; do
@@ -11,26 +13,8 @@ for DIR in "$plugindir"/wp-modules/*; do
 		# Go to the directory of this wp-module.
 		cd "$DIR";
 
-		# If this module does not have a node_modules directory, or if it is empty.
-		if [ ! -d node_modules ] || [ -z "$(ls -A "node_modules")" ]; then
-			# Run npm install for this module.
-			npm install;
-		fi
-
 		# Run the build script for this module.
 		npm run build;
-	fi
-
-	# If this module has a composer.json file.
-	if [ -f "$DIR/composer.json" ];
-	then
-		# Go to the directory of this wp-module.
-		cd "$DIR";
-		
-		# If this module does not have a vendor directory, or if it is empty.
-		if [ ! -d vendor ] || [ -z "$(ls -A "vendor")" ]; then
-			composer install;
-		fi
 	fi
 
 done