+++ phpunit.sh
@@ -2,6 +2,12 @@
 
 # Don't run setup. It's different for phpunit (no node needed, etc).
 # . ./setup.sh
+
+# Note that there are cases where you need node_modules to have been installed for phpunit to pass.
+# For example, wp-scripts will sometimes generate PHP files from JS files, and can cause a run to fail. 
+# For the cases, it is recommended to run install.sh before running phpunit.sh.
+# The reason we don't run npm install here is because we don't mount the node_modules directory into the docker container.
+# This speeds up phpunit tests by a lot, but it means that you need to run install.sh before running phpunit.sh if you need these generated php files.
 while getopts 'p:n:t:f:' flag; do
 	case "${flag}" in
 		p) plugindir=${OPTARG} ;;