+++ test-js.sh
@@ -2,5 +2,6 @@
 
 # Run setup.
 . ./setup.sh
+. ./install-npm.sh
 
 npm run test:js -- --roots "$plugindir"